moved host zone registration tests
